[demo-app] Move TimedErrorMessage to utils
