+++ workspaces/demo-app/src/components/telepath/index.js
@@ -1,2 +1,1 @@
 export { TelepathStatus } from './TelepathStatus'
-export { TimedErrorMessage } from './TimedErrorMessage'
