+++ workspaces/demo-app/src/components/utils/index.js
@@ -1,1 +1,2 @@
 export { TimedStatus } from './TimedStatus'
+export { TimedErrorMessage } from './TimedErrorMessage'
